# Fix spurious perormance degradation introduced in #985